requirements.txt
vkbottle==4.4.6
python-dotenv==1.1.0
SQLAlchemy[asyncio]==2.0.40
aiosqlite==0.21.0